Rename fieldinverter* -> field_inverter*.
