+++ searchlib/src/tests/memoryindex/field_inverter/.gitignore
@@ -1,0 +1,1 @@
+searchlib_field_inverter_test_app
