--- searchlib/src/tests/memoryindex/fieldinverter/.gitignore
@@ -1,1 +1,0 @@
-searchlib_fieldinverter_test_app
