+++ searchlib/src/vespa/searchlib/memoryindex/CMakeLists.txt
@@ -7,7 +7,7 @@ vespa_add_library(searchlib_memoryindex OBJECT
     field_index.cpp
     field_index_collection.cpp
     field_index_remover.cpp
-    fieldinverter.cpp
+    field_inverter.cpp
     memoryindex.cpp
     ordered_field_index_inserter.cpp
     postingiterator.cpp
