+++ searchlib/src/vespa/searchlib/memoryindex/document_inverter.cpp
@@ -2,7 +2,7 @@
 
 #include "document_inverter.h"
 #include "field_index_collection.h"
-#include "fieldinverter.h"
+#include "field_inverter.h"
 #include "ordered_field_index_inserter.h"
 #include "urlfieldinverter.h"
 #include <vespa/document/annotation/alternatespanlist.h>
